+++ import_export/formats/base_formats.py
@@ -202,6 +202,9 @@ class XLSX(TablibFormat):
             dataset.append(row_values)
         return dataset
 
+    def export_data(self, dataset, escape_output=False, **kwargs):
+        return dataset.export(self.get_title(), escape=escape_output, **kwargs)
+
 
 #: These are the default formats for import and export. Whether they can be
 #: used or not is depending on their implementation in the tablib library.